Some fixes: 1) Fixed test configurations and some weight files 2) We decided to make WGraph a child of Graph instead of MWGraph (some methods wouldn't apply). It should expose the usual methods. 3) weight_type is now long double 4) Migrated from MWGraph for bandwidth to WGraph + 2 maps for in/out bandwidth

+++ google_tests/TestGraph.h
@@ -37,7 +37,7 @@ class network_butcher::kfinder::Weighted_Graph<TestGraph<T>,
                                                t_Reversed,
                                                typename TestGraph<T>::Node_Type,
                                                typename TestGraph<T>::Node_Collection_Type,
-                                               double> : base_Weighted_Graph
+                                               long double> : base_Weighted_Graph
 {
 public:
   using Node_Id_Type = std::size_t;
